+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added colour preview/palette selector for colour-based customization options.
 - Added the ability to import 1.0/2.0 .chr files from WoW Model Viewer.
 - Added the ability to import Wowhead Dressing Room characters from URL.
+- Added "Export PNG (3D Preview)" and "Copy to Clipboard (3D Preview)" options to character viewer.
 - Reduced load time of characters tab from 50s to 219ms (230x faster).
 - Removed redundant 'Show M2', 'Show M3', and 'Show WMO' options.
 - Removed 'Show File Data IDs', this is now default behavior.